Convert telemetry float values to doubles across the code base
convert telemetry focused float variables and structures to double variables and structures. This probably breaks the graph and the waypoint panel in some way... Testing will bear it out.

+++ src/Context.java
@@ -262,13 +262,13 @@ public class Context {
     public void sendSetting(int index) {
         settingList.pushSetting(index);
     }
-    public void setSetting(int index, float value) {
+    public void setSetting(int index, double value) {
         settingList.pushSetting(index, value);
     }
-    public void setSettingQuiet(int index, float value) {
+    public void setSettingQuiet(int index, double value) {
         settingList.updateSettingVal(index, value);
     }
-    
+
     public void setTelemetry(int id, float value) {
         telemetry.updateTelemetry(id, (double)value);
     }
